requirements.txt
azure-identity
pandas
python-dotenv